Ajout des controles sur ville

+++ src/main/java/com/sortir/sortir/controller/VilleController.java
@@ -1,9 +1,11 @@
 package com.sortir.sortir.controller;
 
+import com.sortir.sortir.controller.dto.SortieDto;
 import com.sortir.sortir.controller.route.HomeRoute;
 import com.sortir.sortir.controller.route.sortie.SortieAddRoute;
 import com.sortir.sortir.controller.route.ville.VilleEditRoute;
 import com.sortir.sortir.controller.route.ville.VilleRoute;
+import com.sortir.sortir.entity.Sortie;
 import com.sortir.sortir.entity.Ville;
 import com.sortir.sortir.repository.LieuRepository;
 import com.sortir.sortir.repository.ParticipantRepository;
@@ -16,12 +18,15 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.view.RedirectView;
 
 import java.security.Principal;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.ArrayList;
 import java.util.Date;
+import java.util.List;
 
 @Controller
 public class VilleController {
@@ -39,6 +44,9 @@ public class VilleController {
     @Autowired
     ParticipantRepository participantRepository;
 
+
+    List<String> errors = new ArrayList<>();
+
     @GetMapping("/ville/")
     public String add(Model model, Principal principal){
 
@@ -52,7 +60,7 @@ public class VilleController {
     }
 
     @PostMapping("/ville/")
-    public String addValidation(Principal principal, @RequestParam("ville") String villeParam, @RequestParam("codePostal") String codePostal,Model model){
+    public RedirectView addValidation(Principal principal, @RequestParam("ville") String villeParam, @RequestParam("codePostal") String codePostal, Model model, RedirectAttributes ra) throws Exception {
 
         VilleRoute route = new VilleRoute();
         model.addAttribute("route",route);
@@ -61,12 +69,16 @@ public class VilleController {
         ville.setCodePostal(codePostal);
         ville.setVille(villeParam);
 
-        villeRepository.save(ville);
+        if (checkVille(villeParam,codePostal)) {
+            villeRepository.save(ville);
+        } else {
+            ra.addFlashAttribute("errors", errors);
+        }
 
         model.addAttribute("villes", villeRepository.findAll());
         model.addAttribute("user", participantRepository.findByPseudo(principal.getName()));
 
-        return route.getTemplate();
+        return new RedirectView("/ville/");
     }
 
     @PostMapping("/ville/rechercher/")
@@ -94,33 +106,115 @@ public class VilleController {
     }
 
     @PostMapping("/ville/edit/{id}/")
-    public RedirectView editValidation(Principal principal,@PathVariable Integer id,@RequestParam("ville") String villeParam, @RequestParam("codePostal") String codePostal, Model model){
+    public RedirectView editValidation(Principal principal,@PathVariable Integer id,@RequestParam("ville") String villeParam, @RequestParam("codePostal") String codePostal, Model model, RedirectAttributes ra) throws Exception {
 
         VilleRoute route = new VilleRoute();
         model.addAttribute("route",route);
         Ville ville = villeRepository.getOne(id);
         ville.setVille(villeParam);
         ville.setCodePostal(codePostal);
-        villeRepository.save(ville);
+
+        if (checkVille(villeParam,codePostal)) {
+            villeRepository.save(ville);
+        } else {
+            ra.addFlashAttribute("errors", errors);
+        }
+
 
         model.addAttribute("villes", villeRepository.findAll());
         model.addAttribute("user", participantRepository.findByPseudo(principal.getName()));
 
-        return new RedirectView(route.getUrl());
+        return new RedirectView(route.getUrl()+"edit/"+id+"/");
     }
 
     @GetMapping("/ville/delete/{id}/")
-    public String delete(Principal principal,@PathVariable Integer id, Model model) {
+    public RedirectView delete(Principal principal,@PathVariable Integer id, Model model, RedirectAttributes ra) throws Exception {
 
         VilleRoute route = new VilleRoute();
         model.addAttribute("route",route);
 
-        villeRepository.deleteById(id);
+        if (checkIfAttached(villeRepository.getOne(id).getId())) {
+            villeRepository.deleteById(id);
+        } else {
+            ra.addFlashAttribute("errors", errors);
+        }
+
         model.addAttribute("user", participantRepository.findByPseudo(principal.getName()));
 
         model.addAttribute("villes", villeRepository.findAll());
 
-        return route.getTemplate();
+        return new RedirectView("/ville/");
+    }
+
+    public Boolean checkVille(String ville, String codePostal) throws Exception {
+
+        //On vide la liste d'erreur
+        errors.removeAll(errors);
+
+        if (ville == null || ville.isEmpty()) {
+            errors.add("La ville ne peut pas être vide.");
+        }
+        if (codePostal == null || codePostal.isEmpty()) {
+            errors.add("Le code postal ne peut pas être vide.");
+        }
+
+        for(int i=0;i<villeRepository.findAll().size();i++){
+            if(villeRepository.findAll().get(i).getVille().toLowerCase().equals(ville.toLowerCase()) && villeRepository.findAll().get(i).getCodePostal().toLowerCase().equals(codePostal.toLowerCase())){
+                errors.add("Cette ville ("+ville+") existe déjà avec ce code postal ("+codePostal+").");
+            }
+        }
+
+        if (!errors.isEmpty()) {
+            return false;
+        } else {
+            return true;
+        }
+
+    }
+
+    public Boolean checkIfAttached(Integer id) throws Exception {
+
+        //On vide la liste d'erreur
+        errors.removeAll(errors);
+
+        List<SortieDto> sortieList = sortieService.findAll();
+
+        int compteur = 0;
+        int compteur2 = 0;
+
+        for(int i=0;i<sortieList.size();i++){
+
+            if (sortieList.get(i).getLieu().getVille().getId()== id) {
+                compteur = compteur+1;
+            }
+        }
+        for(int i=0;i<participantRepository.findAll().size();i++){
+
+            if (participantRepository.findAll().get(i).getVille().getId()== id) {
+                compteur2 = compteur2+1;
+                System.out.println("compteur2 = "+compteur2);
+            }
+        }
+
+        if(compteur==1){
+            errors.add("Impossible de supprimer cette ville. Elle est liée à " + compteur + " sortie.");
+        }
+        if(compteur>1){
+            errors.add("Impossible de supprimer cette ville. Elle est liée à " + compteur + " sorties.");
+        }
+        if(compteur2==1){
+            errors.add("Impossible de supprimer cette ville. Elle est liée à " + compteur2 + " participant.");
+        }
+        if(compteur2>1){
+            errors.add("Impossible de supprimer cette ville. Elle est liée à " + compteur2 + " participants.");
+        }
+
+        if (!errors.isEmpty()) {
+            return false;
+        } else {
+            return true;
+        }
+
     }
